scan_gitcode/requirements.txt
PyYAML==6.0.1
pyetherpadlite==1.1.1
click==8.1.7
requests==2.32.3
pandas==2.2.2
Jinja2==3.1.4
